[WIP] Graph metrics improvements
