+++ metrics/web/lib/common/presentation/graphs/colored_bar.dart
@@ -2,19 +2,24 @@ import 'package:flutter/material.dart';
 
 /// Rectangle bar of the [BarGraph] with painted in [color].
 class ColoredBar extends StatelessWidget {
+  /// The color of this bar.
   final Color color;
+
+  /// The radius of the border of this bar.
   final BorderRadiusGeometry borderRadius;
+
+  /// The border decoration of this bar.
   final BoxBorder border;
+
+  /// The padding to inset this bar.
   final EdgeInsets padding;
+
+  /// The width of this bar.
   final double width;
 
   /// Creates the [ColoredBar].
-  ///
-  /// [padding] is the padding to inset this bar.
-  /// [border] is the border decoration of this bar.
-  /// [borderRadius] is the radius of the border of this bar.
-  /// [color] is the color of this bar.
-  /// [width] is the width of the bar.
+  /// 
+  /// The [padding] defaults to the [EdgeInsets.zero].
   const ColoredBar({
     Key key,
     this.padding = EdgeInsets.zero,
