+++ metrics/web/lib/dashboard/presentation/state/project_metrics_notifier.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:math';
 
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:metrics/dashboard/domain/entities/collections/date_time_set.dart';
 import 'package:metrics/dashboard/domain/entities/metrics/build_result_metric.dart';
@@ -10,8 +10,10 @@ import 'package:metrics/dashboard/domain/entities/metrics/performance_metric.dar
 import 'package:metrics/dashboard/domain/usecases/parameters/project_id_param.dart';
 import 'package:metrics/dashboard/domain/usecases/receive_project_metrics_updates.dart';
 import 'package:metrics/dashboard/domain/usecases/receive_project_updates.dart';
-import 'package:metrics/dashboard/presentation/model/build_result_bar_data.dart';
 import 'package:metrics/dashboard/presentation/model/project_metrics_data.dart';
+import 'package:metrics/dashboard/presentation/view_models/build_result_metric_view_model.dart';
+import 'package:metrics/dashboard/presentation/view_models/build_result_view_model.dart';
+import 'package:metrics/dashboard/presentation/view_models/performance_metric_view_model.dart';
 import 'package:metrics_core/metrics_core.dart';
 
 /// The [ChangeNotifier] that holds the projects metrics state.
@@ -166,15 +168,12 @@ class ProjectMetricsNotifier extends ChangeNotifier {
     final buildResultMetrics = _getBuildResultMetrics(
       dashboardMetrics.buildResultMetrics,
     );
-    final averageBuildDuration =
-        dashboardMetrics.performanceMetrics.averageBuildDuration.inMinutes;
     final numberOfBuilds = dashboardMetrics.buildNumberMetrics.numberOfBuilds;
 
     projectsMetrics[projectId] = projectMetrics.copyWith(
       performanceMetrics: performanceMetrics,
       buildResultMetrics: buildResultMetrics,
       buildNumberMetric: numberOfBuilds,
-      averageBuildDurationInMinutes: averageBuildDuration,
       coverage: dashboardMetrics.coverage,
       stability: dashboardMetrics.stability,
     );
@@ -184,36 +183,45 @@ class ProjectMetricsNotifier extends ChangeNotifier {
   }
 
   /// Creates the project performance metrics from [PerformanceMetric].
-  List<Point<int>> _getPerformanceMetrics(PerformanceMetric metric) {
+  PerformanceMetricViewModel _getPerformanceMetrics(PerformanceMetric metric) {
     final performanceMetrics = metric?.buildsPerformance ?? DateTimeSet();
 
     if (performanceMetrics.isEmpty) {
-      return [];
+      return const PerformanceMetricViewModel();
     }
 
-    return performanceMetrics.map((metric) {
+    final performance = performanceMetrics.map((metric) {
       return Point(
         metric.date.millisecondsSinceEpoch,
         metric.duration.inMilliseconds,
       );
     }).toList();
+
+    final averageBuildDuration = metric.averageBuildDuration.inMinutes;
+
+    return PerformanceMetricViewModel(
+      performance: performance,
+      value: averageBuildDuration,
+    );
   }
 
   /// Creates the project build result metrics from [BuildResultMetric].
-  List<BuildResultBarData> _getBuildResultMetrics(BuildResultMetric metrics) {
+  BuildResultMetricViewModel _getBuildResultMetrics(BuildResultMetric metrics) {
     final buildResults = metrics?.buildResults ?? [];
 
     if (buildResults.isEmpty) {
-      return [];
+      return const BuildResultMetricViewModel();
     }
 
-    return buildResults.map((result) {
-      return BuildResultBarData(
+    final buildResultViewModels = buildResults.map((result) {
+      return BuildResultViewModel(
         url: result.url,
         buildStatus: result.buildStatus,
         value: result.duration.inMilliseconds,
       );
     }).toList();
+
+    return BuildResultMetricViewModel(buildResults: buildResultViewModels);
   }
 
   /// Cancels all created subscriptions.
